Cleaned up the AgentSystem implementation.
