+++ agents_playground/agents/spec/byproduct_definition.py
@@ -1,0 +1,5 @@
+from typing import NamedTuple
+
+class ByproductDefinition(NamedTuple):
+  name: str
+  type: type
