+++ agents_playground/agents/systems/agent_auditory_system.py
@@ -12,6 +12,6 @@ class AgentAuditorySystem(AgentSystem):
     self.name = 'auditory-system'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """What does the agent hear?"""
     return
