+++ agents_playground/agents/systems/agent_gustatory_system.py
@@ -11,7 +11,7 @@ class AgentGustatorySystem(AgentSystem):
     self.name = 'gustatory-system'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """
     - Is there anything in the agent's mouth? What does it taste like? 
     """
