+++ agents_playground/agents/systems/agent_olfactory_system.py
@@ -11,7 +11,7 @@ class AgentOlfactorySystem(AgentSystem):
     self.name = 'olfactory-system'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """
     - What does the agent smell? 
     """
