+++ agents_playground/agents/systems/agent_perception_system.py
@@ -15,13 +15,13 @@ class AgentPerceptionSystem(AgentSystem):
     self.name = 'agent-perception'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """
     TODO: Collect all sensory information that the agent is experiencing.
     """
     return
   
-  def after_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _after_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """
     TODO: Prepare the stimuli for the attention subsystem.
     - This could mean storing it in one of the Agent's memory stores.
