+++ agents_playground/agents/systems/agent_somatosensory_system.py
@@ -11,7 +11,7 @@ class AgentSomatosensorySystem(AgentSystem):
     self.name = 'somatosensory-system'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """
     - What is the agent touching? 
     - What is their temperature? Are they hot, cold?
