+++ agents_playground/agents/systems/agent_vestibular_system.py
@@ -11,6 +11,6 @@ class AgentVestibularSystem(AgentSystem):
     self.name = 'vestibular-system'
     self.subsystems = SimpleNamespace()
 
-  def before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
+  def _before_subsystems_processed(self, characteristics: AgentCharacteristics, agent_phase: AgentLifeCyclePhase) -> None:
     """What is impacting the agent's balance? Are they nauseous?"""
     return
